Flatten packages in minerva database package, and rename to make things clearer.
