+++ app/src/main/java/be/ugent/zeus/hydra/data/database/Database.java
@@ -5,9 +5,9 @@ import android.arch.persistence.room.RoomDatabase;
 import android.arch.persistence.room.TypeConverters;
 import android.content.Context;
 
-import be.ugent.zeus.hydra.data.database.minerva.agenda.AgendaDao;
-import be.ugent.zeus.hydra.data.database.minerva.announcement.AnnouncementDao;
-import be.ugent.zeus.hydra.data.database.minerva.course.CourseDao;
+import be.ugent.zeus.hydra.data.database.minerva.AgendaDao;
+import be.ugent.zeus.hydra.data.database.minerva.AnnouncementDao;
+import be.ugent.zeus.hydra.data.database.minerva.CourseDao;
 import be.ugent.zeus.hydra.data.database.migrations.*;
 import be.ugent.zeus.hydra.data.dto.DateTypeConverters;
 import be.ugent.zeus.hydra.data.dto.minerva.AgendaItemDTO;
