+++ app/src/main/java/be/ugent/zeus/hydra/data/database/RepositoryFactory.java
@@ -3,9 +3,9 @@ package be.ugent.zeus.hydra.data.database;
 import android.content.Context;
 import android.support.annotation.NonNull;
 
-import be.ugent.zeus.hydra.data.database.minerva.agenda.DatabaseAgendaItemRepository;
-import be.ugent.zeus.hydra.data.database.minerva.announcement.DatabaseAnnouncementRepository;
-import be.ugent.zeus.hydra.data.database.minerva.course.DatabaseCourseRepository;
+import be.ugent.zeus.hydra.data.database.minerva.AgendaDatabaseRepository;
+import be.ugent.zeus.hydra.data.database.minerva.AnnouncementDatabaseRepository;
+import be.ugent.zeus.hydra.data.database.minerva.CourseDatabaseRepository;
 import be.ugent.zeus.hydra.data.dto.minerva.AgendaMapper;
 import be.ugent.zeus.hydra.data.dto.minerva.AnnouncementMapper;
 import be.ugent.zeus.hydra.data.dto.minerva.CourseMapper;
@@ -28,7 +28,7 @@ public class RepositoryFactory {
     public static synchronized AgendaItemRepository getAgendaItemRepository(Context context) {
         if (agendaItemRepository == null) {
             Database database = Database.get(context);
-            agendaItemRepository = new DatabaseAgendaItemRepository(database.getAgendaDao(), CourseMapper.INSTANCE, AgendaMapper.INSTANCE);
+            agendaItemRepository = new AgendaDatabaseRepository(database.getAgendaDao(), CourseMapper.INSTANCE, AgendaMapper.INSTANCE);
         }
         return agendaItemRepository;
     }
@@ -37,7 +37,7 @@ public class RepositoryFactory {
     public static synchronized AnnouncementRepository getAnnouncementRepository(Context context) {
         if (announcementRepository == null) {
             Database database = Database.get(context);
-            announcementRepository = new DatabaseAnnouncementRepository(database.getAnnouncementDao(), CourseMapper.INSTANCE, AnnouncementMapper.INSTANCE);
+            announcementRepository = new AnnouncementDatabaseRepository(database.getAnnouncementDao(), CourseMapper.INSTANCE, AnnouncementMapper.INSTANCE);
         }
         return announcementRepository;
     }
@@ -46,7 +46,7 @@ public class RepositoryFactory {
     public static synchronized CourseRepository getCourseRepository(Context context) {
         if (courseRepository == null) {
             Database database = Database.get(context);
-            courseRepository = new DatabaseCourseRepository(database.getCourseDao(), CourseMapper.INSTANCE);
+            courseRepository = new CourseDatabaseRepository(database.getCourseDao(), CourseMapper.INSTANCE);
         }
         return courseRepository;
     }
