+++ app/src/test/java/be/ugent/zeus/hydra/data/database/minerva/mocks/MockAnnouncementDao.java
@@ -5,7 +5,7 @@ import android.database.sqlite.SQLiteConstraintException;
 import android.os.Build;
 import android.support.annotation.RequiresApi;
 
-import be.ugent.zeus.hydra.data.database.minerva.announcement.AnnouncementDao;
+import be.ugent.zeus.hydra.data.database.minerva.AnnouncementDao;
 import be.ugent.zeus.hydra.data.dto.minerva.AnnouncementDTO;
 import be.ugent.zeus.hydra.data.dto.minerva.CourseDTO;
